GFA

Types

PanGraph.Graphs.GFA.Path — Type

struct Path
    name     :: String
    segments :: Array{Node,1}
    circular :: Bool
end

Store a GFA path, i.e. a sequence of segments that represents an observed genome.

PanGraph.Graphs.GFA.Segment — Type

struct Segment
    name     :: String
    sequence :: Array{UInt8}
    depth    :: Int
end

Store a GFA segment, i.e. an edge of an alignment graph that holds a contiguous sequence Depth is the number of genomes, including duplications, that contain the sequence.

Functions

PanGraph.Graphs.marshal_gfa — Method

marshal_gfa(io::IO, G::Graph; opt=nothing)

Output pangraph G to IO stream io. opt can include two functions, to be accessed in fields connect and output. connect is a function that takes a node and returns true or false if it should be connected in the GFA output. output is an equivalent function signature, but controls whether the node is output at all.