Nodes

Types

PanGraph.Graphs.Nodes.Node — Type

mutable struct Node{T}
	block  :: T
	strand :: Bool
end

Node represents a portion of a sequence path that passes through a single block. strand stores whether we pass along the forward strand of block (if true) or reverse (if false).

PanGraph.Graphs.Nodes.Node — Method

Node{T}(b::T; strand=true)

Create a Node that passed through block b. Default to forward strand orientation.

Functions

Base.length — Method

length(n::Node) = length(n.block, n)

Return the length of sequence stored within node n