PanGraph globals

Types

PanGraph.Alignment — Type

mutable struct Alignment{T <: Union{String,Nothing,Array{Tuple{Int,Char}}}}
	qry::Hit
	ref::Hit
	matches::Int
	length::Int
	quality::Int
	orientation::Bool
	cigar::T
	divergence::Union{Float64,Nothing}
	align::Union{Float64,Nothing}
end

Alignment is a pairwise homologous alignment between two sequences.

PanGraph.Hit — Type

mutable struct Hit
	name::String
	length::Int
	start::Int
	stop::Int
	seq::Maybe{Array{UInt8,1}}
end

Hit is one side of a pairwise alignment between homologous sequences.

PanGraph.PanContigs — Type

struct PanContigs
	name     :: T
	sequence :: T
end

A synonym for a consensus sequence of Block.

Functions